EmbeddedClientFactory now creates embedded database with isPersitent=false by default.
When operating in non-persistent mode, the file system is still checked for stores, thus
giving the impression that we're able to operate in read-only mode, which is not really
the case, as while the Store objects will be loaded, there will be no pager able to
load the persisted indexes and pages. (this should be tidied up soon).

This can be overridden, and is used just for PersistentTest for now.  This is done by
calling EmbeddedClientFactory.setPesistent(true) before calling getInstance().

+++ com.wwm.db.server/src/main/java/com/wwm/db/internal/server/ServerStore.java
@@ -54,12 +54,18 @@ public class ServerStore implements Serializable {
 	 */
 	private long savedDbVersion = 0; 
 	
+	/**
+	 * 
+	 * @param storageBasePath - disk location, or null if no storage is available
+	 */
 	public ServerStore(String storageBasePath, String storeName, int storeId) {
 		this.storeName = storeName;
 		this.storeId = storeId;
 		
-		String diskName = makeDirForNewStore(storeName, storageBasePath);
-		this.path = storageBasePath + File.separatorChar + diskName;
+		if (storageBasePath != null) {
+			String diskName = makeDirForNewStore(storeName, storageBasePath);
+			this.path = storageBasePath + File.separatorChar + diskName;
+		}
 		
 		namespaces = new Namespaces(this);
 		pagerClassTokenCache = new ClassTokenCache(true);
@@ -130,6 +136,8 @@ public class ServerStore implements Serializable {
 	}
 
 	public boolean deletePersistentData() {
+		if (path==null) return true; // nothing to delete
+		
 		if (!namespaces.deletePersistentData()) return false;
 		if (!deleteTempDir()) return false;
